Move Payments -> PaymentProtocol
