+++ examples/bip70.fetch.php
@@ -2,8 +2,8 @@
 
 require "../vendor/autoload.php";
 
-use BitWasp\Bitcoin\Payments\PaymentRequestBuilder;
-use BitWasp\Bitcoin\Payments\PaymentRequestSigner;
+use BitWasp\Bitcoin\PaymentProtocol\PaymentRequestBuilder;
+use BitWasp\Bitcoin\PaymentProtocol\PaymentRequestSigner;
 
 $time = $_GET['time'];
